split kind service in main and post services
